implemented all points up to 5th
